Kore.Variables.UnifiedVariable: refreshElementVariable, refreshSetVariable
